+++ test/Sentry.AspNetCore.Tests/VersioningTests.PreFlightIgnoresTransaction.verified.txt
@@ -20,7 +20,7 @@
       }
     }
   },
-  payloads: [
+  Payloads: [
     {
       Source: {
         Name: Unknown Route,
